hw_2
distribution test for C fix
